feat: implement VendingMachine interactions

+++ java-vendingmachine-precourse/src/main/java/vendingmachine/scenario/OutputView.java
@@ -22,7 +22,7 @@ final class OutputView {
         }
     }
 
-    void printUserBalanceInProgress(int balance) {
+    void printUserBalanceInProgress(long balance) {
         console.printLine(String.format("투입 금액: %s원", balance));
     }
